requirements_mobile.txt
streamlit==1.28.0
opencv-python==4.8.1.78
pytesseract==0.3.10
gspread==5.12.0
google-auth==2.23.4
pillow==10.1.0
numpy==1.24.3
requests==2.31.0